+++ client/scripts/db/hurricane.js
@@ -171,7 +171,7 @@ module.exports = {
     ],
   },
   memoryPosition: [
-    { offset: 0x323, description: 'current text', type: 'string' },
+    { offset: 0x326, description: 'current text', type: 'string' },
     { offset: 0x3AC, description: 'current player', type: 'uint8' },
     { offset: 0x3AD, description: 'current ball', type: 'uint8' },
     { offset: 0x188D, description: 'player count', type: 'uint8' },